チャンネル番号でドラムトラック判定するように変更
