+++ SFZ_MIDI_Player/source/SampleSource.cpp
@@ -3,6 +3,7 @@
 #include <AudioLoadManager.hpp>
 #include <SampleSource.hpp>
 #include <SamplePlayer.hpp>
+#include <AudioStreamRenderer.hpp>
 
 double Envelope::level(double noteOnTime, double noteOffTime, double time) const
 {
